pyproject.toml
[project]
name = "shackbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "async-unzip>=0.5.5",
    "discord-py>=2.6.4",
    "media-toolkit>=0.2.22",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
shackbot = "shackbot.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
shackbot = ["static/*"]

[tool.uv]
package = true
exclude-dependencies = ["onnxruntime-gpu"]
